Fixing overlap in mt utils and edp
